# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Projet API",
    "description": "Proof-of-work hiring infrastructure.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/programmes/{programme_id}/submissions": {
      "get": {
        "tags": [
          "judging"
        ],
        "summary": "List Submission Cards",
        "description": "FR-1051 - the judging day list, in the order people pitch.\n\nEvery non-excluded seat appears here, including people who declined\nrecruitment sharing. That consent only gates the candidate pool later.",
        "operationId": "list_submission_cards_programmes__programme_id__submissions_get",
        "parameters": [
          {
            "name": "programme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Programme Id"
            }
          },
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionCard"
                  },
                  "title": "Response List Submission Cards Programmes  Programme Id  Submissions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/programmes/{programme_id}/participants/{participant_id}/score": {
      "get": {
        "tags": [
          "judging"
        ],
        "summary": "Get Scoring Card",
        "operationId": "get_scoring_card_programmes__programme_id__participants__participant_id__score_get",
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Participant Id"
            }
          },
          {
            "name": "programme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Programme Id"
            }
          },
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringCard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "judging"
        ],
        "summary": "Update Scoring Card",
        "description": "Auto-saving, like the prescreen rubric. A judge is watching a pitch, not\nfilling in a form, so nothing waits on a Save they will forget to press.",
        "operationId": "update_scoring_card_programmes__programme_id__participants__participant_id__score_patch",
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Participant Id"
            }
          },
          {
            "name": "programme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Programme Id"
            }
          },
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoringUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringCard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/programmes/{programme_id}/participants/{participant_id}/testimonial/draft": {
      "post": {
        "tags": [
          "judging"
        ],
        "summary": "Draft Testimonial Endpoint",
        "description": "Draft from endorsed skills and the brief. Does not save or publish.\n\nThe model only rearranges facts already on the card. A generated paragraph\nis still the company's word, so it lands in the textarea rather than on\nthe profile.",
        "operationId": "draft_testimonial_endpoint_programmes__programme_id__participants__participant_id__testimonial_draft_post",
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Participant Id"
            }
          },
          {
            "name": "programme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Programme Id"
            }
          },
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestimonialDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/programmes/{programme_id}/close": {
      "post": {
        "tags": [
          "judging"
        ],
        "summary": "Close",
        "description": "Turn the evening's scores into the thing participants keep.\n\nSafe to repeat: a judge who finishes their card late still reaches the\nprofile on the next close.",
        "operationId": "close_programmes__programme_id__close_post",
        "parameters": [
          {
            "name": "programme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Programme Id"
            }
          },
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloseoutOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SubmissionCard": {
        "properties": {
          "participant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Participant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "organisation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organisation"
          },
          "run_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Order"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked"
          },
          "complete": {
            "type": "boolean",
            "title": "Complete"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/SubmittedLink"
            },
            "type": "array",
            "title": "Links"
          },
          "scored": {
            "type": "boolean",
            "title": "Scored"
          },
          "your_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Your Total"
          }
        },
        "type": "object",
        "required": [
          "participant_id",
          "name",
          "organisation",
          "run_order",
          "status",
          "submitted_at",
          "locked",
          "complete",
          "links",
          "scored",
          "your_total"
        ],
        "title": "SubmissionCard",
        "description": "One participant, what they handed in, and where their scoring stands."
      },
...
    }
  }
}